requirements.txt
Django==1.5.1
Pillow==2.1.0
South==0.8.1
Twisted==13.1.0
Whoosh==2.4.1
argparse==1.2.1
autobahn==0.5.14
distribute==0.6.24
django-bootstrap-toolkit==2.14.0
django-extensions==1.1.1
django-haystack==2.0.0
django-hstore==1.1.1
django-image-cropping==0.6.4
djorm-ext-core==0.4.2
djorm-ext-expressions==0.4.4
djorm-ext-pgarray==0.5
easy-thumbnails==1.1
httplib2==0.8
ipython==0.13.2
pep8==1.4.6
psycopg2==2.5.1
pyflakes==0.7.3
python-memcached==1.53
redis==2.7.6
six==1.3.0
uWSGI==1.9.13
wsgiref==0.1.2
zope.interface==4.0.5
